% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmanData.R
\docType{data}
\name{Infinium5}
\alias{Infinium5}
\alias{lvr.combat}
\alias{lvr.harman}
\alias{md.combat}
\alias{md.harman}
\title{Infinium Methylation BeadChip batch correction reference data}
\format{
Four matrices each with 899255 rows and 5 columns. One row per CpG
site across the 450K and EPIC designs and one column for each of the
reference datasets. \code{lvr.combat} and \code{lvr.harman} are the matrices
of log variance ratio (LVR) statistics for ComBat and Harman, respectively.
While, \code{md.combat} and \code{md.harman} are the matrices of mean
differences post-batch correction.
}
\usage{
data(Infinium5)
}
\value{
\code{\link{promise}}
}
\description{
Summary statistics after batch-effect correction per CpG probe
across 5 Infinium Methylation datasets.
}
\examples{
data(Infinium5)
}
\references{
... awaiting publication
}
