% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmanData.R
\docType{data}
\name{NPM}
\alias{NPM}
\alias{npm.data}
\alias{npm.info}
\title{Non-pregant mice study data}
\format{
Two data frames. \code{npm.data} has 35,512 probesets (rows) and 24
biological samples (columns). While \code{npm.info} is a description of the
samples, with two columns:
\describe{
  \item{\code{Treatment}}{the treatment applied to the mice}
  \item{\code{Batch}}{batch processing number}
}
}
\usage{
data(NPM)
}
\value{
\code{\link{promise}}
}
\description{
An experiment to test skin penetration of metal oxide
nanoparticles following topical application of sunscreens. The data comprises
three treatment groups plus a control group, with six replicates in each
group, making a total of 24 Affymetrix MoGene 1.0 ST arrays. There were a
total of three processing batches of eight arrays, each consisting of 2
replicates per group. Arrays were normalised and background adjusted as a
whole using the RMA procedure in MATLAB.
}
\examples{
data(NPM)
}
\references{
Osmond-McLeod et al. Nanotoxicology (2014).
doi: 10.3109/17435390.2013.855832
}
