% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmanData.R
\docType{data}
\name{Infinium5}
\alias{Infinium5}
\alias{lvr.combat}
\alias{lvr.harman}
\alias{md.combat}
\alias{md.harman}
\title{Infinium Methylation BeadChip batch correction reference data}
\format{
Four matrices each with 899255 rows and 5 columns. One row per CpG
site across the 450K and EPIC designs and one column for each of the
reference datasets. The matrices contain log variance ratio (LVR) statistics
for ComBat and Harman and mean differences post-batch correction:
\describe{
  \item{\code{lvr.combat}}{LVR for ComBat}
  \item{\code{lvr.harman}}{LVR for Harman}
  \item{\code{md.combat}}{mean differences for ComBat}
  \item{\code{md.harman}}{mean differences for Harman}
}
}
\usage{
data(Infinium5)
}
\value{
\code{\link{promise}}
}
\description{
Summary statistics after batch-effect correction per CpG probe
across 5 Infinium Methylation datasets.
}
\examples{
data(Infinium5)
}
\references{
... awaiting publication
}
